Update deploy command to use controller container
