--- src/api/controller.ts
@@ -1,4 +1,0 @@
-export * from './app';
-export * from './artifact';
-export * from './deployment';
-export * from './log';
