--- src/api/index.ts
@@ -1,3 +1,0 @@
-import * as controller from './controller';
-
-export { controller };
