+++ src/types/index.ts
@@ -1,10 +1,6 @@
 export * from './app';
 export * from './appConfig';
-export * from './artifact';
 export * from './cliContext';
 export * from './clusterConfig';
-export * from './deployment';
-export * from './deploymentItemConfig';
 export * from './globalConfig';
-export * from './logEntry';
 export * from './tenantConfig';
